% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddTransaction.R
\name{AddTransaction}
\alias{AddTransaction}
\alias{addTransaction}
\alias{editTransaction}
\alias{viewLedgerMonth}
\title{Add, edit or view transactions}
\usage{
addTransaction(ledger = viewLedger(file = file),
  file = viewLedgerFile(), description = NULL, budget = NULL,
  date = NULL, account = NULL, amount = NULL)

editTransaction(ledger = viewLedger(file = file),
  file = viewLedgerFile(), ...)

viewLedgerMonth(ledger = viewLedger(file = file),
  file = viewLedgerFile())
}
\arguments{
\item{ledger}{Data frame ledger. The ledger where each row is a transaction
and with columns for \code{year}, \code{month}, \code{day},
\code{description}, \code{budget}, and one column for each
financial account. All columns should be numeric except for
\code{description} and \code{budget}.}

\item{file}{The path for the ledger for loading and saving. Has an underlying
default that needs to be changed in the \code{ViewLedger} file.}

\item{descriptions}{Values in the descriptions column to search for. Uses
regular expressions.}

\item{budget.categories}{Values from the factor column of budget category to
search for. Must be exact.}

\item{amount.ops, account.amounts}{Operators to search transactions. For
example, to find transactions where more than one thousand dollars were
spent, use \code{amount.ops = `<`, account.amounts = -1000}.}

\item{account.categories}{Restricts the transactions shown to ones where the
account(s) in this argument are not 0.}

\item{query}{Asks the user for descriptions, transaction amounts, etc.
manually. Queried values will overwrite the input arguments. Default
behavior is to query if all the input options are \code{NULL}.}
}
\description{
Add a transaction to the ledger via inputs to the console with
  \code{addTransaction}.\cr \code{editTransaction} will display the
  transactions for the given year, month, day. It defaults to show the
  current month's. If \code{yr}, \code{mo}, \code{dy} is \code{NULL}, it will
  show transactions from all years, months, days, respectively, within the
  others. It also takes other arguments to narrow the options to show. These
  options are the same as for \code{viewLedger}, because
  \code{editTransaction} will call the former to limit the options shown. \cr
  \code{viewLedgerMonth} is a convenient wrapper to show the transactions for
  just the current month. Instead, the functions with date subdivisions are
  more appropriate. This is the appropriate way to view a reasonable number
  of transactions at one time. It takes the \code{year}, \code{month}, and
  \code{day} to prune down the ledger. If any date specification is
  \code{NULL}, it allows all values for that specific date column.
  \code{year} and \code{month} default to \code{currentYear()} and
  \code{currentMonth()}, while \code{day} defaults to all values via
  \code{NULL}. The ledger can further be restricted by including an early or
  late bound with \code{from} and \code{to}. These require a \code{Date}
  type. You can use \code{componentsToDate} to build one using a year, month,
  and day. Both methods can be done alone or in tandem.\cr For a transaction
  to be shown, it has to fit ALL of the input argument criteria. However,
  each of the input arguments can be a single value or can be a vector. If
  one is a vector, a transaction that fits any of the values in the vector
  will be shown.
}
