% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bofa.R
\name{bofa_credit}
\alias{bofa_credit}
\title{Classify new transaction for Bank oF America accounts.}
\usage{
bofa_credit(new, skrooge, output_path = "BankOfAmerica.csv",
  after_date = today())
}
\arguments{
\item{new}{Path to the CSV transaction file.}

\item{skrooge}{Path to the Skrooge database file.}

\item{output_path}{Path to the output CSV file.}

\item{after_date}{Classify only transactions younger than this date.}
}
\description{
Classify new transaction for Bank oF America accounts.
}
