% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify_transaction}
\alias{classify_transaction}
\title{find which known transactions are most similar to those not yet
classified (new)}
\usage{
classify_transaction(new, known)
}
\description{
find which known transactions are most similar to those not yet
classified (new)
}
