% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFinances.R
\name{.binByDate}
\alias{.binByDate}
\title{Bins the date column in a ledger data frame by a given timeframe. Returns a tibble grouped by (only) the timeframe}
\usage{
.binByDate(ledger, by = "month")
}
\description{
Bins the date column in a ledger data frame by a given timeframe. Returns a tibble grouped by (only) the timeframe
}
