% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFinances.R
\name{.binFun}
\alias{.binFun}
\title{Performs a supplied function on the amount column of a ledger tibble, likely grouped by some time frame in the date column.}
\usage{
.binFun(ledger.tbl, fun)
}
\description{
Performs a supplied function on the amount column of a ledger tibble, likely grouped by some time frame in the date column.
}
