% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFinances.R
\name{.plotBudgetsDate}
\alias{.plotBudgetsDate}
\title{Gets a reduced ledger for easier plotting based on a given month and certain specifications}
\usage{
.plotBudgetsDate(ledger, yr, mo, budget.categories)
}
\description{
Gets a reduced ledger for easier plotting based on a given month and certain specifications
}
