% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcu.R
\name{fcu_any}
\alias{fcu_any}
\title{Classify new transaction for FCU accounts.}
\usage{
fcu_any(new, skrooge, output_path = "FCU.csv", after_date = today())
}
\arguments{
\item{new}{Path to the CSV transaction file.}

\item{skrooge}{Path to the Skrooge database file.}

\item{output_path}{Path to the output CSV file.}

\item{after_date}{Classify only transactions younger than this date.}
}
\description{
Classify new transaction for FCU accounts.
}
