% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finances-package.R
\docType{package}
\name{finances-package}
\alias{finances}
\alias{finances-package}
\title{finances: Luckett Finances}
\description{
A Shiny application that tracks monthly expenses, utilities, and savings contributions. Will also track progress toward retirement.
}
\author{
\strong{Maintainer}: Crystal Luckett \email{cnguyen6292@gmail.com}

Authors:
\itemize{
  \item Daniel Luckett \email{daniel.j.luckett@gmail.com}
}

}
\keyword{internal}
