% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaEdits.R
\name{MetaEdits}
\alias{MetaEdits}
\alias{createLedger}
\alias{addAccount}
\alias{setBudgets}
\title{High-level edits}
\usage{
createLedger()

addAccount(ledger = viewLedger(file = file), file = viewLedgerFile())

setBudgets(ledger = viewLedger(file = file), file = viewLedgerFile())
}
\arguments{
\item{ledger}{Data frame ledger. The ledger where each row is a transaction
and with columns for \code{year}, \code{month}, \code{day},
\code{description}, \code{budget}, and one column for each
financial account. All columns should be numeric except for
\code{description} and \code{budget}.}

\item{file}{The path for the ledger for loading and saving. Has an underlying
default that needs to be changed in the \code{ViewLedger} file.}
}
\description{
These functions have the power to add new accounts, create new
  ledgers, etc.
}
\details{
\code{createLedger} builds a new ledger in the standard form. It
  queries the user for account names, initial values, budget categories and
  amounts, and file location.\cr \code{addAccount} can take a ledger and add
  a new account name to it. It queries the user for the account name and the
  column position in the ledger.
}
