% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFinances.R
\name{PlotFinances}
\alias{PlotFinances}
\alias{plotAccountsChange}
\alias{plotAccountsCumulative}
\alias{plotAssetsChange}
\alias{plotAssetsCumulative}
\alias{plotBudgetsHistory}
\alias{plotBudgetsBar}
\alias{plotBudgetsPie}
\title{Plot financial data}
\usage{
plotAccountsChange(ledger = viewLedger(file = file),
  file = viewLedgerFile(), from = NULL, to = NULL)

plotAccountsCumulative(ledger = viewLedger(file = file),
  file = viewLedgerFile(), from = NULL, to = NULL)

plotAssetsChange(ledger = viewLedger(file = file), file = viewLedgerFile(),
  from = NULL, to = NULL)

plotAssetsCumulative(ledger = viewLedger(file = file),
  file = viewLedgerFile(), from = NULL, to = NULL)

plotBudgetsHistory(ledger = viewLedger(file = file),
  file = viewLedgerFile(), budget.categories = viewBudgetCategoriesAll(),
  from = NULL, to = NULL)

plotBudgetsBar(ledger = viewLedger(file = file), file = viewLedgerFile(),
  yr = currentYear(), mo = currentMonth(),
  budget.categories = viewBudgetCategoriesAll())

plotBudgetsPie(ledger = viewLedger(file = file), file = viewLedgerFile(),
  yr = currentYear(), mo = currentMonth(),
  budget.categories = viewBudgetCategoriesSpendingSmall())
}
\arguments{
\item{ledger}{Data frame ledger. The ledger where each row is a transaction
and with columns for \code{year}, \code{month}, \code{day},
\code{description}, \code{budget.category}, and one column for each
financial account. All columns should be numeric except for
\code{description} and \code{budget.category}.}

\item{file}{The path for the ledger for loading and saving. Has an underlying
default that needs to be changed in the \code{ViewLedger} file.}

\item{from}{Boundaries on the ledger dates. Requires Date class.}

\item{to}{Boundaries on the ledger dates. Requires Date class.}

\item{yr}{Year, month, day. When one of these is \code{NULL}, function
behavior is to allow all possible values.}

\item{mo}{Year, month, day. When one of these is \code{NULL}, function
behavior is to allow all possible values.}

\item{budget.category}{A character vector of budget category names.}
}
\value{
\code{ggplot} object, which is automatically printed if not stored in a variable.
}
\description{
Various plotting functions to show assets, accounts, budgets, etc. over different time periods.
}
\details{
All \code{Plot} functions show the dollar amounts over some
  specified time. \code{Change} shows the monthly adjustment to the ledger
  (i.e., income and spending per month) while \code{Cumulative} shows the
  total amounts up until that month. \cr\code{Assets} combines all accounts,
  while \code{Accounts} splits each financial account up for viewing. \cr The
  \code{Budget} series shows amounts for each budget category. \code{History}
  shows all months in the ledger, while \code{plotBudgetsBar} allows the user
  to input a specific month and give breakdowns of spending compared to the
  set budget values. This defaults to the current month. This takes a
  character vector of budget category names. This can be a custom list such
  as \code{c("Entertainment", "Food")}, or you can use the package functions
  \code{viewBudgetCategoriesAll()}, \code{viewBudgetCategoriesSpending()}, or
  \code{viewBudgetCategoriesReal()}. \code{plotBudgetsPie} combines all the
  budgets and shows the amount spent and remaining next to the days left in
  the month for easy visual tracking of progress.
}
