% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViewDate.R
\name{ViewDate}
\alias{ViewDate}
\alias{currentDay}
\alias{currentMonth}
\alias{currentYear}
\alias{yesterday}
\alias{lastWeek}
\alias{lastMonth}
\alias{lastYear}
\alias{componentsToDate}
\title{Get date info}
\usage{
currentDay()

currentMonth()

currentYear()

yesterday()

lastWeek()

lastMonth()

lastYear()

componentsToDate(yr = currentYear(), mo = currentMonth(),
  dy = currentDay())
}
\value{
Each of the \code{current*} functions return an integer. \code{last*}
  and \code{componentsToDate} return a \code{Date} class.
}
\description{
The \code{current} series gets today's year, month, or day,
  returning just an integer. The \code{last} series (including
  \code{yesterday}) returns a full \code{Date} class for a recent date for
  easily providing values to the \code{from} input argument in other
  functions. For custom dates, use \code{componentsToDate}. To return today's
  date in the proper format, call \code{componentsToDate()}.
}
